fixed bsq

+++ bsq/bsq.h
@@ -6,7 +6,7 @@
 /*   By: ibaby <ibaby@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/07 01:06:28 by ibaby             #+#    #+#             */
-/*   Updated: 2024/12/07 19:09:06 by ibaby            ###   ########.fr       */
+/*   Updated: 2024/12/07 23:40:21 by ibaby            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,23 +24,23 @@
 # define IS_NOT !=
 
 typedef struct s_obstacle {
-	size_t		x;
-	size_t		y;
+	int		x;
+	int		y;
 	struct	s_obstacle	*next;
 }	t_obstacle;
 
 typedef struct s_square {
-	size_t		x;
-	size_t		y;
-	size_t		size;
+	int		x;
+	int		y;
+	int		size;
 }	t_square;
 
 typedef struct s_data {
 	t_square	square;
 	t_obstacle	*obstacles;
 	char		**map;
-	size_t		line_len;
-	size_t		lines;
+	int		line_len;
+	int		lines;
 	char		empty;
 	char		full;
 	char		obstacle;
